componentizing tour to better compartmentalize functionality. implemented a isLoading state to the toursegments so they don't skip over tour steps on accident
