+++ frontend/src/components/dwe/cameras/camera-controls.tsx
@@ -10,7 +10,7 @@ import {
   MonitorCog,
 } from "lucide-react";
 
-import { TOUR_STEP_IDS } from "@/components/tour/tour-lib/tour-constants";
+import { TOUR_STEP_IDS } from "@/components/tour/tour-constants";
 import {
   Accordion,
   AccordionContent,
