+++ frontend/src/components/dwe/cameras/device-card.tsx
@@ -1,4 +1,4 @@
-import { TOUR_STEP_IDS } from "@/components/tour/tour-lib/tour-constants";
+import { TOUR_STEP_IDS } from "@/components/tour/tour-constants";
 import {
   Card,
   CardContent,
