+++ frontend/src/components/dwe/log-page/log-viewer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { API_CLIENT } from "@/api";
-import { TOUR_STEP_IDS } from "@/components/tour/tour-lib/tour-constants";
+import { TOUR_STEP_IDS } from "@/components/tour/tour-constants";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
