+++ frontend/src/components/themes/mode-toggle.tsx
@@ -3,7 +3,7 @@ import { Moon, Sun, SunMoon } from "lucide-react";
 import { Button } from "@/components/ui/button";
 
 import { useTheme } from "@/components/themes/theme-provider";
-import { TOUR_STEP_IDS } from "@/components/tour/tour-lib/tour-constants";
+import { TOUR_STEP_IDS } from "@/components/tour/tour-constants";
 
 export function ModeToggle() {
   const { setTheme, theme } = useTheme();
