+++ frontend/src/components/tour/index.ts
@@ -1,0 +1,4 @@
+export { TOUR_STEP_IDS } from "./tour-constants";
+export { useTour } from "./tour-context";
+export { TourDialog } from "./tour-overlay";
+export { TourProvider } from "./tour-provider";
